feat(rest-client): Complete the demo version of request sending and response receiving

+++ restclient-core/src/main/java/esa/restclient/core/RestClientBuilder.java
@@ -190,7 +190,7 @@ public class RestClientBuilder implements Reusable<RestClientBuilder>, RestClien
     }
 
     @Override
-    public long readTimeout() {
+    public int readTimeout() {
         return httpClientBuilder.readTimeout();
     }
